Added exception handling

+++ src/se/sme/rest/IntervalResource.java
@@ -2,6 +2,7 @@ package se.sme.rest;
 
 import java.util.List;
 
+import javax.ejb.EJBTransactionRolledbackException;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
@@ -14,6 +15,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.Response;
 
 import se.sme.domain.IntervalInstance;
 import se.sme.service.TimeManagementService;
@@ -33,35 +35,60 @@ public class IntervalResource {
     @GET
     @Produces("application/JSON")
     @Path("{startNo}")
-    public List<IntervalInstance> getIntervalByStart(@PathParam("startNo") int hour) {
-        return service.searchByStart(hour);
+    public Response getIntervalByStart(@PathParam("startNo") int hour) {
+        if (hour < 0) {
+            return Response.status(400).build();
+        } //bad request
+        List<IntervalInstance> list = service.searchByStart(hour);
+        if(list.size() == 0) { //not found
+            return Response.status(404).build();
+        }
+        return Response.ok(list).build();
     }
 
     @POST
     @Produces("application/JSON")
     @Consumes("application/JSON")
-    public IntervalInstance createInterval(IntervalInstance interval) {
-        service.registerInterval(interval);
-        return interval;
+    public Response createInterval(IntervalInstance interval) {
+        IntervalInstance ii;
+        try {
+            ii = new IntervalInstance(interval.getStartHour(), interval.getEndHour());
+        } catch (IllegalArgumentException ex) {
+          //bad request
+            return Response.status(400).build();
+        }
+        service.registerInterval(ii);
+        return Response.ok(interval).build();
     }
 
     @DELETE
     @Path("{hour}")
-    public void delete(@PathParam("hour") int hour) {
+    public Response delete(@PathParam("hour") int hour) {
+        if(hour < 0) {
+            return Response.status(400).build(); //bad
+        }
         List<IntervalInstance> li = service.searchByStart(hour);
+        if(li.size() < 1) {
+            return Response.status(404).build(); //not found
+        }
+        System.out.println(li.size());
         for (IntervalInstance i : li) {
             service.delete(i);
         }
-        // return Response.ok().build();
+        return Response.ok().build();
     }
 
     @PUT
-    public void update(@DefaultValue("1") @QueryParam("hour") Integer hour, @QueryParam("id") Integer id) {
-        if (hour >= 0 && id != null) {
-            // return Response.ok().build();
-            service.addHours(hour, id);
+    public Response update(@DefaultValue("1") @QueryParam("hour") Integer hour, @QueryParam("id") Integer id) {
+        if(id < 1 || hour < 0) {
+            return Response.status(400).build(); //bad
         }
-        // return Response.status(400).build();
+        try {
+            service.addHours(hour, id);
+            return Response.ok().build();
+        } catch (EJBTransactionRolledbackException ex) {
+            return Response.status(404).build(); //not found
+        }
     }
 
 }
